external_pc_ui/requirements.txt
eclipse-zenoh>=1.0.0
msgpack>=1.0.0
PyQt5>=5.15.0
pyqtgraph>=0.13.0
PyOpenGL>=3.1.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
